add synthesize simulator
